+++ FE/src/components/Loader.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function Loader() {
+  return <div>반찬이 없어요 😊...</div>;
+}
+
+export default Loader;
